fix package.json so you can run etherpad-lite installed by npm
This patch adds all files/dirs that are needed to run etherpad-lite
without using any of the hacky bin/ scripts.

Since the top level dir is not able to contribute files, we move the
settings.json.template as an example to src/example/.

This is step one that is validated to support global installs.